separate tests into basic and ones that hit the db
